+++ kotti_newsitem/__init__.py
@@ -3,9 +3,12 @@
 """
 Created on 2013-02-07
 :author: Andreas Kaiser (disko)
+:contributors: Jeff Pittman (geojeff)
 """
 
 from pyramid.i18n import TranslationStringFactory
+
+from kotti.resources import Document
 
 from kotti.views.slots import assign_slot
 
@@ -15,7 +18,14 @@ _ = TranslationStringFactory('kotti_newsitem')
 def kotti_configure(settings):
 
     settings['pyramid.includes'] += ' kotti_newsitem'
+
     settings['kotti.available_types'] += ' kotti_newsitem.resources.NewsItem'
+
+    if 'kotti_newsitem.num_news' in settings:
+        settings['kotti_newsitem.num_news'] = int(
+            settings['kotti_newsitem.num_news'])
+    else:
+        settings['kotti_newsitem.num_news'] = 10
 
     if 'kotti_newsitem.widget.num_news' in settings:
         settings['kotti_newsitem.widget.num_news'] = int(
@@ -28,6 +38,9 @@ def kotti_configure(settings):
 
 
 def includeme(config):
+
+    Document.type_info.add_selectable_default_view("news_listing",
+                                                   _("News Listing"))
 
     config.add_translation_dirs('kotti_newsitem:locale')
     config.scan(__name__)