+++ CHANGES.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+0.2
+---
+
+-   Added a full-page style news listing.
+
 0.1
 ---